Drawer functionality
A drawer component was created for when the user wishes to focus on the profile of a specific tenant. The drawer opens when the user clicks on the 'View' button next to a tenant's name on the TenantTable component. The user can then quickly browse all relevant information about the tenant and close the drawer by clicking the X button on the top right. In the next step the ability to edit tenant information on the drawer will be added.