Pop up question on Add Bill Payment Form
Added a pop-up question on the Add Bill Payment form so that the date a bill was paid is requested only if the bill payment has been paid. The submit button was also centred at the middle of the page. Finally the Input Adornment was tweaked so that it appears only once a user starts typing in the amount field.