Finalized status and column filters; began refactoring process
This commit finalises the implementation of status and column filters. Users now have the ability to clear filters easily by selecting an X button that appears upon making a choice. Additionally, the usability of column dropdowns has been improved to enhance user interaction. The handleRentStatusChange, handleBillStatusChange, and handleTaskStatusChange functions were consolidated into a single handleStatusChange function to streamline the management of status filters. As part of efforts to improve maintainability, a StatusFilter.jsx file was introduced containing a reusable StatusFilter component. This component standardises the implementation of status filters throughout the application, although it has not yet replaced existing filters in OrderTable.jsx. This commit sets the foundation for future enhancements in filter management and overall code refinement.