Added security and support page
Added security features such as express-mongo-sanitize, helmet, rateLimit, and HTML sanitization via yup to enhance the application's security. Additionally, I introduced a support page to assist users with any questions. Lastly, I included a disclaimer to remind users not to enter real data, as this is an open project accessible to everyone.